requirements.txt
json
selenium
mysql-connector-python
